+++ lightly_studio_view/src/lib/components/index.ts
@@ -48,6 +48,7 @@ export { default as VideoSampleMetadata } from '$lib/components/VideoSampleMetad
 export { default as GridHeader } from '$lib/components/GridHeader/GridHeader.svelte';
 export { default as DatasetGridHeader } from '$lib/components/DatasetGridHeader/DatasetGridHeader.svelte';
 export { default as SidePanelTabs } from '$lib/components/SidePanelTabs/SidePanelTabs.svelte';
+export { default as FilterPanel } from '$lib/components/FilterPanel/FilterPanel.svelte';
 export { default as CaptionsItem } from '$lib/components/Captions/CaptionsItem/CaptionsItem.svelte';
 export { default as Typography } from '$lib/components/Typography/Typography.svelte';
 export { default as GroupsGrid } from '$lib/components/GroupsGrid/GroupsGrid.svelte';